% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma_function.R
\name{ewma_function}
\alias{ewma_function}
\title{Exponentially Weighted Moving Average Function}
\usage{
ewma_function(data, lambda = 0.01, cs = 0.5, sigma = 0.5, L = 1.5)
}
\arguments{
\item{data}{The dataframe you want to run the EWMA for, must include binary responses in a column named "correct" and reaction times in a column named "RT"}

\item{lambda}{Weight of previously calculated EWMA statistics. Between 0 and 1. If 1, only the current observation is used. If 0, all previous observations are weighted equally. Defaults to .01}

\item{cs}{Control mean, used as the previous EWMA statistic for the first iteration. Defaults to .5 (chance)}

\item{sigma}{Control standard deviation. Defaults to .5}

\item{L}{Width of control limits in standard deviations. Low to ensure a sensitive test. Defaults to 1.5}
}
\value{
List containing the generated cutoff threshold and a dataframe of EWMA statistics and UCLs for each trial with their corresponding reaction times.
}
\description{
Calculates the EWMA statistic and upper control limit (UCL) for each trial in a dataframe in order of reaction time. It also generates a cutoff threshold past which accuracy should be above chance.
}
\keyword{ddm}
\keyword{ewma}
\keyword{hddm}

