% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewma_graph.R
\name{ewma_graph}
\alias{ewma_graph}
\title{Exponentially Weighted Moving Average Graph Function}
\usage{
ewma_graph(ewma_results)
}
\arguments{
\item{ewma_results}{List that must contain a cutoff reaction time called "cutoff" and a dataframe, called "results," with a column of EWMA statistics called "cs", a column of UCLs called "UCL," and a column of reaction times called "rt."}
}
\value{
Lineplot of EWMA statistics and UCLs by reaction time. 
ewma_graph()
}
\description{
This function takes the results of an EWMA filter (such as the list returned by ewma_function) and generates a line graph of EWMA statistic and upper control limit (UCL) by reaction time.
}
\keyword{ddm}
\keyword{ewma}
\keyword{graph}
\keyword{hddm}

